separate Cart ViewModel from Cart View Controller
